+++ crates/scuttlecast/src/receiver/sink.rs
@@ -5,7 +5,7 @@ use derive_more::Constructor;
 use crate::receiver::reorderer::Reorderer;
 
 pub trait Sink {
-    async fn write(&mut self, block_no: u32, offset: u64, bytes: &[u8]) -> io::Result<()>;
+    async fn write(&mut self, block_no: u64, offset: u64, bytes: &[u8]) -> io::Result<()>;
     async fn finish(&mut self) -> io::Result<()>;
 }
 
@@ -14,7 +14,7 @@ pub struct FileSink {
     file: std::fs::File,
 }
 impl Sink for FileSink {
-    async fn write(&mut self, _block_no: u32, offset: u64, bytes: &[u8]) -> io::Result<()> {
+    async fn write(&mut self, _block_no: u64, offset: u64, bytes: &[u8]) -> io::Result<()> {
         self.file.write_all_at(bytes, offset)
     }
 
@@ -28,7 +28,7 @@ pub struct StreamSink {
     reorderer: Reorderer,
 }
 impl Sink for StreamSink {
-    async fn write(&mut self, block_no: u32, _offset: u64, bytes: &[u8]) -> io::Result<()> {
+    async fn write(&mut self, block_no: u64, _offset: u64, bytes: &[u8]) -> io::Result<()> {
         self.reorderer.on_block(block_no, bytes.to_vec()).await;
         Ok(())
     }
